+++ backend/package.json
@@ -89,6 +89,7 @@
     "Nestjs",
     "Typescript",
     "GraphQL",
-    "Nestjs GraphQL"
+    "Nestjs GraphQL",
+    "Agile"
   ]
 }
